app/requirements.txt
flask
flask_sqlalchemy
flask_rest_jsonapi
flask_jsontools
